Logout and dynamic admin check
Moved the function that handle loggin in to GitHub into action.js. Also wrote a function to handle logging out.
Initialized the session object within the NavBar component and passed it as a prop to the links component.
Since links has the useState hook it can be an async function. Changed the isAdmin to make it dynamic, it checks the session object to identify the same.