Register page created
Moved login page into (auth) folder to keep all the authentication related pages together.
Created a register page and added some styling to it.
Wrote a function for the registration form in action.js which saves the registered account in the database, it checks if the user already exists first.