Registration page message display
Modified the register function to return the error object.
Created a separate component for the registration form, since it is a client side component.
Using the useFormState hook the error messages are dynamically displayed.
Issue with the useEffect hook. after registration is a success the login page isn't automatically redirected into